middle(0405)

+++ src/main/java/com/cgv/s1/oproduct/OproductDTO.java
@@ -13,7 +13,8 @@ public class OproductDTO {
 	private String productName;
 	private String productDetail;
 	private Integer productPrice;
-	private Double productDC;
+	//퍼센트 double에서 integer로 바꿈
+	private Integer productDC;
 	private Integer productStock;
 	private Integer productSale;
 	private Date regDate;
@@ -85,10 +86,10 @@ public class OproductDTO {
 	public void setProductPrice(Integer productPrice) {
 		this.productPrice = productPrice;
 	}
-	public Double getProductDC() {
+	public Integer getProductDC() {
 		return productDC;
 	}
-	public void setProductDC(Double productDC) {
+	public void setProductDC(Integer productDC) {
 		this.productDC = productDC;
 	}
 	public Integer getProductStock() {
